removed some warnings

+++ Application/Source/MeshBuilder.cpp
@@ -13,7 +13,7 @@ Mesh* MeshBuilder::Generatering(const std::string& meshName, float lengthX, floa
 	std::vector<Vertex> vertex_buffer_data;
 	std::vector<unsigned>index_buffer_data;
 	for (int theta = 0; theta <= 360; theta += 10) {
-		v.pos.Set(10 * cos(theta * (PI / 180)), 0, 10 * sin(theta * (PI / 180))); 
+		v.pos.Set(10.f * float(cos(theta * (PI / 180)))), 0.f, 10.f * sin(theta * (PI / 180)); 
 		v.normal.Set(0, 1, 0);
 		v.color.Set(Red); 
 		vertex_buffer_data.push_back(v);
@@ -171,7 +171,7 @@ Mesh* MeshBuilder::GenerateCylinder(const std::string& meshName, float lengthX, 
 			vertex_buffer_data.push_back(v);
 			v.pos.Set(1 * cos(theta * Rad), (-height / 2) + (stack+1) * stackheight, 1 * sin(theta * Rad));
 			v.color.Set(name);
-			v.normal.Set(cos(theta * Rad), 0, sin(theta * Rad));
+			v.normal.Set(float(cos(theta * Rad)), 0, float(sin(theta * Rad)));
 			vertex_buffer_data.push_back(v);
 			index += 2;
 		}
@@ -192,7 +192,7 @@ Mesh* MeshBuilder::GenerateCylinder(const std::string& meshName, float lengthX, 
 		v.color.Set(name);
 		v.normal.Set(0, -1, 0);
 		vertex_buffer_data.push_back(v);
-		v.pos.Set(1 * cos(theta * Rad), -height / 2,1 * sin(theta * Rad));
+		v.pos.Set(1 * float(cos(theta * Rad)), -height / 2,1 * float(sin(theta * Rad)));
 		v.color.Set(name);
 		v.normal.Set(0, -1, 0);
 		vertex_buffer_data.push_back(v);
@@ -221,7 +221,7 @@ Mesh* MeshBuilder::GenerateCone(const std::string& meshName, float lengthX, floa
 	int index = 0;
 	Mesh* mesh = new Mesh(meshName);
 	for (int theta = 0; theta <= 360; theta += 10) {
-		v.pos.Set(height / height*1 * cos(theta * Rad), -height / 2, height / height *1* sin(theta * Rad));
+		v.pos.Set(height / height*1 * float(cos(theta * Rad)), -height / 2, height / height *1* float(sin(theta * Rad)));
 		v.color.Set(name);
 		v.normal.Set(height * (height / height * 1 * cos(theta * Rad)), 1, height * (height / height * 1 * sin(theta * Rad)));
 		vertex_buffer_data.push_back(v);
@@ -236,7 +236,7 @@ Mesh* MeshBuilder::GenerateCone(const std::string& meshName, float lengthX, floa
 		v.color.Set(name);
 		v.normal.Set(0, -1, 0);
 		vertex_buffer_data.push_back(v);
-		v.pos.Set(height / height * 1 * cos(theta * Rad), -height / 2, height / height * 1 * sin(theta * Rad));
+		v.pos.Set(height / height * 1 * float(cos(theta * Rad)), -height / 2, height / height * 1 * float(sin(theta * Rad)));
 		v.color.Set(name);
 		v.normal.Set(0, -1, 0);
 		vertex_buffer_data.push_back(v);
@@ -433,12 +433,12 @@ Mesh* MeshBuilder::GenerateColliderBox(const std::string& meshName, double legnt
 	//sides
 	v.pos.Set(float(-legnthx) / 2, float(lengthy / 2), float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
 	v.pos.Set(float(-legnthx) / 2, float(-lengthy / 2), float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, lengthy / 2, float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, -lengthy / 2, float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(-legnthx) / 2, lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(-legnthx) / 2, -lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
-	v.pos.Set(float(legnthx) / 2, -lengthy / 2, float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, float(lengthy / 2), float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, float(-lengthy / 2), float(-lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(-legnthx) / 2, float(lengthy / 2), float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(-legnthx) / 2, float(-lengthy / 2), float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, float(lengthy / 2), float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
+	v.pos.Set(float(legnthx) / 2, float(-lengthy / 2), float(lengthz / 2)); v.color.Set(Red); vertex_buffer_data.push_back(v);
 	Mesh* mesh = new Mesh(meshName);
 	for (int i = 0; i < 24; i++) {
 		index_buffer_data.push_back(i);
@@ -461,7 +461,7 @@ Mesh* MeshBuilder::GenerateCircle(const std::string& meshName, float lengthX, fl
 	std::vector<Vertex> vertex_buffer_data;
 	std::vector<unsigned>index_buffer_data;
 	for (int theta = 0; theta <= 360; theta+=10) {
-		v.pos.Set(1*cos(theta*(PI/180)), 0, 1*sin(theta*(PI/180))); v.color.Set(Red); vertex_buffer_data.push_back(v);
+		v.pos.Set(1*cos(theta*(PI/180)), 0, 1*float(sin(theta*(PI/180)))); v.color.Set(Red); vertex_buffer_data.push_back(v);
 		v.pos.Set(0, 0, 0);v.color.Set(Blue);vertex_buffer_data.push_back(v);
 	}
 	for (int i = 0; i < 74; i++) {
@@ -484,7 +484,7 @@ Mesh* MeshBuilder::GenerateSemiCircle(const std::string& meshName, float lengthX
 	std::vector<Vertex> vertex_buffer_data;
 	std::vector<unsigned>index_buffer_data;
 	for (int theta = 0; theta <= 180; theta += 10) {
-		v.pos.Set(1 * cos(theta * (PI / 180)), 0, 1 * sin(theta * (PI / 180))); v.color.Set(color); vertex_buffer_data.push_back(v);
+		v.pos.Set(1 * float(cos(theta * (PI / 180))), 0, 1 * float(sin(theta * (PI / 180)))); v.color.Set(color); vertex_buffer_data.push_back(v);
 		v.pos.Set(0, 0, 0); v.color.Set(color); vertex_buffer_data.push_back(v);
 	}
 	for (int i = 0; i < 38; i++) {
